refactor(taiko-client-rs): merge tx_list into codec, rename api_impl to handlers
- Move tx_list constants and decompress_tx_list into codec.rs, delete tx_list.rs
- Rename api/service/api_impl.rs to handlers.rs
- Move ensure_fee_recipient_allowed into handlers.rs, delete lookahead.rs
- Group all constants at the top of codec.rs
