+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/api/service/payload_build.rs
@@ -1,5 +1,7 @@
 //! Payload build/signing helpers used by `build_preconf_block`.
 
+use crate::codec::decompress_tx_list;
+
 use super::*;
 
 impl<P> WhitelistApiService<P>
@@ -13,7 +15,7 @@ where
         prev_randao: B256,
         signature: [u8; 65],
     ) -> Result<TaikoPayloadAttributes> {
-        let tx_list = crate::tx_list::decompress_tx_list(request.transactions.as_ref())?;
+        let tx_list = decompress_tx_list(request.transactions.as_ref())?;
 
         let block_metadata = TaikoBlockMetadata {
             beneficiary: request.fee_recipient,
