+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/lib.rs
@@ -12,7 +12,6 @@ mod network;
 mod operator_set;
 mod preconf_ingress_sync;
 mod runner;
-mod tx_list;
 
 pub use error::{Result, WhitelistPreconfirmationDriverError};
 pub use metrics::WhitelistPreconfirmationDriverMetrics;
